renamed Contract to Contracts
