+++ src/Symfony/Component/HttpKernel/DataCollector/DataCollectorInterface.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\HttpKernel\DataCollector;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Contract\Service\ResetInterface;
+use Symfony\Contracts\Service\ResetInterface;
 
 /**
  * DataCollectorInterface.
