+++ src/Symfony/Component/Validator/Validator/TraceableValidator.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\Validator\Validator;
 
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
-use Symfony\Contract\Service\ResetInterface;
+use Symfony\Contracts\Service\ResetInterface;
 
 /**
  * Collects some data about validator calls.
