+++ src/Symfony/Contracts/.gitignore
