+++ src/Symfony/Contracts/CHANGELOG.md
